Generate two-sample t-test results for multiple sub-groups

This function produce two-sample t-test (two-tailed with confident interval at 0.95) results for multiple sub-groups and provides with a nice output in a table format. It can also add adjusted p values for multiple comparison issue.

t_test_two_sample(data, x, y, paired = FALSE, p_adjust = "bonferroni")

Arguments

data
A grouped data frame. It should be grouped by the intended sub-groups which you want to do the same t-test.

x
Column name of the variable which contains data values that you want to test (see t.test and details).

y
Column name of the variable which contains data values of group assignments for the test values (see t.test and details).

paired
a logical indicating whether you want a paired t-test.

p_adjust
character indicating which method should be used for adjusting multiple comparisons (see p.adjust and details). The default "bonferroni" corresponds to Bonferroni adjustment.

Value

A data.frame with the t-statistics table consisting of characters. The columns that are always present are: group variable(s), tvalue, df (degrees of freedom), p, and p_adjustmethod(s).

Examples

t_test_two_sample(color_index_two_sample, x = "color_effect", y = "group", paired = TRUE)

#> # A tibble: 8 x 5
#> # Groups:   roi_id [8]
#>   roi_id  tvalue    df      p p_bonferroni
#>   <chr>    <dbl> <dbl>  <dbl>        <dbl>
#> 1 AnG      0.570    28 0.573        1     
#> 2 dLatIPS  1.68     28 0.104        0.835 
#> 3 LO       2.27     28 0.0311       0.249 
#> 4 pIPS     1.85     28 0.0752       0.601 
#> 5 V1       1.22     28 0.232        1     
#> 6 vIPS     2.67     28 0.0124       0.0991
#> 7 vLatIPS  1.69     28 0.101        0.811 
#> 8 VTC      1.16     28 0.257        1     

# use bonferroni and fdr method for adjusted p values.
library(magrittr)
color_index_two_sample %>%
  t_test_two_sample(
  x = "color_effect", y = "group", paired = TRUE, p_adjust = c("bonferroni","fdr")
  )

#> # A tibble: 8 x 6
#> # Groups:   roi_id [8]
#>   roi_id  tvalue    df      p p_bonferroni  p_fdr
#>   <chr>    <dbl> <dbl>  <dbl>        <dbl>  <dbl>
#> 1 AnG      0.570    28 0.573        1      0.573 
#> 2 dLatIPS  1.68     28 0.104        0.835  0.167 
#> 3 LO       2.27     28 0.0311       0.249  0.124 
#> 4 pIPS     1.85     28 0.0752       0.601  0.167 
#> 5 V1       1.22     28 0.232        1      0.294 
#> 6 vIPS     2.67     28 0.0124       0.0991 0.0991
#> 7 vLatIPS  1.69     28 0.101        0.811  0.167 
#> 8 VTC      1.16     28 0.257        1      0.294